Refactoring specification level design responsibility
