+++ preferred/app/Application/Console/Kernel.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Preferred\Application\Kernels;
+namespace Preferred\Application\Console;
 
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
 
-class Console extends ConsoleKernel
+class Kernel extends ConsoleKernel
 {
     protected $commands = [];
