+++ preferred/app/Application/Http/Middlewares/RedirectIfAuthenticated.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Preferred\Application\Middlewares;
+namespace Preferred\Application\Http\Middlewares;
 
 use Closure;
 use Preferred\Interfaces\Http\Controllers\ResponseTrait;
