+++ preferred/app/Application/Http/Middlewares/TrimStrings.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Preferred\Application\Middlewares;
+namespace Preferred\Application\Http\Middlewares;
 
 use Illuminate\Foundation\Http\Middleware\TrimStrings as Middleware;
