+++ preferred/app/Application/Http/Middlewares/TrustProxies.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Preferred\Application\Middlewares;
+namespace Preferred\Application\Http\Middlewares;
 
 use Fideloper\Proxy\TrustProxies as Middleware;
 use Illuminate\Http\Request;
